add population & sex columns to sex marker output
If you use the option to pull sex-identifying from your dataset, then the 'Population ID' column will now be printed in that excel file output. Additionally, support for an optional 'Sex' column is added which can contain phenotypic sex data. This is also printed to the sex-identifying marker output (if used).